inventory/anode-system/scraper/requirements.txt
playwright==1.54.0
supabase==2.10.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
requests==2.32.3
Pillow==11.0.0
cryptography==44.0.0
schedule==1.2.2
colorama==0.4.6
tqdm==4.67.1